qcacld-3.0: Add IPA SW route control for seamless roaming scenario
If running tx traffic with IPA path during roaming on the seamless
roaming scenario(all the external APs work on the same gateway),
as the IPA does not know whether the STA is roaming or not, and
the IPA also does not have the flow control for the tx data transfer,
and then the tx packets will be dropped by wlan fw if IPA continue to
send packets to wlan fw during the roaming between roam start and
roam completed.

To avoid the tx packets drop in this case, need to switch the IPA
(HW) path to SW path(TCP/IP Stack from kernel) when roam start,
and switch back to the IPA(HW) path after roam completed.

When roam start, host will stop the kernel tx queue before
switching the IPA path to SW path, and then tx packets will be queued
in kernel after switched to SW path.

After roam completed, host will restart the tx queue and switch the
SW path to IPA path.
Note that There is not conflict even if the tx packets sent to wlan
fw from SW path and IPA path in parallel.

Change-Id: I3aaeafd8135992e50d3cb4fa54a65492dda682c4
CRs-Fixed: 3805471