qcacld-3.0: Add support to set latency level for WFC_STATE cmd
Currently, the legacy vendor command
QCA_WLAN_VENDOR_ATTR_CONFIG_LATENCY_LEVEL is used to set the
latency level for the WLAN latency module in HOST/FW. The value
of the latency level varies from NORMAL (0) to ULTRALOW (4).

As per the new requirement, when the host receives the
QCA_WLAN_VENDOR_ATTR_CONFIG_WFC_STATE vendor command to set
and reset the WFC state, the host should behave as follows:

On receiving WFC state = 1, the host first caches the latency
level configured via the legacy vendor command for the same
client ID, if any, and then sends the WMI_WLM_CONFIG_CMDID
command to the FW with the latency level value = 1.

On receiving WFC state = 0, the host sends the cached latency
level value to the FW.

Change-Id: I8e9a0a0a21cea3cedb83fa1ad1dc267db953e56e
CRs-Fixed: 4065408