qcacld-3.0: Fix invalid channel width change detection logic
While updating the channel width based on beacon, if AP
advertises 80 + 80, we allow channel switch to 80 + 80 instead
of restricting it to 80 MHz operation. The current operating
channel width is freshly calculated instead of using the
current PE session channel width. This results in doing
CSA after connection with wrong secondary channel value.

As in the STA mode we don't support 80 + 80 operation,
restrict the channel width to 80 MHz if the AP is advertising
80+80 MHz operation in VHT operation IE. Fetch the current
operating channel width from PE session instead of calculating
again.

CRs-Fixed: 4046941
Change-Id: Ib8c5dece6404492f361892adfb512a59c2f252da