qcacld-3.0: Check for VLP cut-off freq before early termination
During ACS scan, if the device finds clean channel, it terminates
ACS scan early and selects the channel. Currently this channel is
not checked against VLP cut-off frequency whether is below to
cut-off or not. This results in ACS not deprioritizing 6 GHz freq
which are below to cut-off frequency.

Add check to validate if the chan falls in cut-off range, then
deprioritize the chan so that weight for these chan is calculated
high.

Change-Id: I49b24081ac3cc7f3c785131062f95dd3b7763a55
CRs-Fixed: 4046062
(cherry picked from commit d7d211164abc3a2d001af28635628162b48948b9)