qcacld-3.0: Do not reset num_pkts of fisa flow
Currently, when flow's average pkt count is calculated and
the flow is not active for 1s, the num_pkts value which has
current total number of packets is being reset to
num_pkt_prev which contains packet count used during
last calculation. This is causing packet counts to get
missed when flow is being checked for flow classification.
To fix this, set num_pkts_prev to num_pkts so that latest
values are picked during both average calculation and
packet rate.

Change-Id: I5d0e7bbe280f94e6df5e51358245f542278eac98
CRs-Fixed: 4067281