qcacld-3.0: SAP Restart on Unsafe Channel in 5 GHz Band
SAP is attempting to restart on the unsafe channel 5745, despite the
entire 5 GHz band being marked as unsafe. Since 5745 is a NAN social
channel, it is being removed from the unsafe list and marked as safe.
This applies only if NAN discovery is enabled.

Verify if NAN discovery is disabled, and if it is, do not mark 5745 as a
safe channel.

CRs-Fixed: 4071752
Change-Id: I274eaee88d182760d201dc72d885be2051066682