qcacld-3.0: Define link_info's link flags with qdf_bitmap
Currently, the link_info's link flag is defined as an
unsigned long, and kernel APIs (set_bit, test_bit, clear_bit)
are used to manipulate it. However, these APIs might expect
an unsigned long array, and a static analyzer tool has
flagged this as an out-of-bounds (OOB) access.

To address this, use QDF APIs to define and operate on the
bitmap, ensuring proper handling.

Change-Id: I0d2275c7ad0dc578c0fdfefe488d7d0dc351cad5
CRs-Fixed: 3870137