qcacld-3.0: Do not enable/disable active APF when DHCP is in progress
Currently active apf is enabled/disabled when setsuspend mode is called
with idle_monitor 1 and 0 respectively. This setsuspend mode is also
called when DHCP is started and stopped and active APF is
enabled/disabled when DHCP is started/stopped which is not the expected
behavior.

To resolve this issue, don't enable/disable active APF when setsuspend
is called during DHCP phase.

CRs-Fixed: 4070984
Change-Id: I7e18efe18e417531f78f0276f4c91aa5efebf002