qcacld-3.0: Use hash table for Tx flow tracking
As sk_txtime_unused variable is a 6 bit value, only 64
flows were being stored, this was causing lots of collisions
and active flows get missed in tracking.
To fix this, create a hash table of size 64 and store the
index of hash table in sk_txtime_unused.
Update flow eviction logic to not evict if flow is selected
for sampling or is classified.

Change-Id: Ibb8adc9334ffe4b682f48078d34634e928559905
CRs-Fixed: 4090489