qcacld-3.0: Add scan rest time and burst duration for RM scan request
When initiating scan request while processing the beacon report
request frame, scan rest time and burst duration is not populated
in the scan request parameter sent to the scan command to
Firmware. This results in the traffic latency when there is
significant number of channel to be scanned as part of processing
beacon report request frames. This latency is observed as the
scan rest time is set to '0' due to which the firmware
does not return to home channel during the scan process.

Add scan rest time and burst duration in scan request parameter
of beacon report scan request in order to allow the firmware
to return to the home channel to reduce the traffic latency.

Change-Id: I9f0aac69afa35625eb977e99e2c2f4f816dbb38b
CRs-Fixed: 4079515