qcacld-3.0: Add ini and interface support for roam optimization
Add an ini to enable the MLO roam optimization with
partner link bringup offloaded to the host driver.

Also, intersect the firmware support for the feature
with the ini config and send the updated config to
the firmware via RSO START command. This allows the
firmware to decide whether to decide the path for
roaming.

Change-Id: Ie76affbdf03b1e54a12f8c846790c6540f42ef99
CRs-Fixed: 4101599