qcacld-3.0: Fix potential out-of-bounds access in memdump_read
Currently, the host does not check the pos value when driverdump
is called for the first time. This can lead to an out-of-bounds
access issue when filling the user space buffer if the pos value
is greater than the driver_dump_size.

To address this, return an error if the pos value is greater or
equal to the driver dump size after driver_dump_size is calculated.

CRs-Fixed: 4082374
Change-Id: I974786da6389c27c7d11bf49fe3682e403f54697
(cherry picked from commit f92c26c77bcb0d6c1be853def474e88eb6cf9a7b)