qcacld-3.0: Do not send enable APF command in disconnected state
Currently, the host sends the enable active mode APF command to the
firmware in a disconnected state when the screen turns off,
which can cause unexpected firmware behavior.

To resolve this issue, avoid sending the enable APF command if the
screen turns off while the STA is in a disconnected state.
Additionally, if the STA connects to an AP while the screen is off,
implement logic to send the enable active mode APF command to the
firmware after the connection is established.

Change-Id: I2a8203dd34c3bff8fa4b3ea3ee97cd3736375ff4
CRs-Fixed: 4069146