qcacld-3.0: Pick up new frequency if CSA target channel not valid
In the SAP + STA concurrency case:
1. SAP + STA running on DFS channel
2. external AP move to other 5 GHz channel that is not valid for DUT
(DUT may trim the channel list to exclude that channel, but this
channel is under regulatory for external AP)
3. WLAN FW report the first CSAIE received event to host
4. SAP will try to move to the target frequency

Since target channel is not valid for DUT, so channel switch
failure.

The fix is to select one new valid channel if target channel
is invalid for this case to make SAP CSA successfully.

Change-Id: Ic92285940ce16842f99a37530616e3f150e7dde9
CRs-Fixed: 4107544