qcacld-3.0: Avoid dropping the CSA event due to bandwidth mismatch
To avoid CAC wait, for example, by starting the AP on BW 160 MHz but
limiting traffic to BW 80 MHz, the channel width of some AP's beacons
may not match the value in the association response.

This change prevents the DUT from dropping the CSA event, as the new
channel width is the same as the channel width in the pe_session.

Change-Id: Ia23697ab81f9a89f2a788b760b32a658fda794e5
CRs-Fixed: 4103645