qcacld-3.0: Fix SCC issue in SAP+LL_LT_SAP
In XPAN, SCC is not allowed with SAP and
LL_LT_SAP when both are operating on 5 GHz.
Initially, SAP starts on 5 GHz and LL_LT_SAP
also starts on 5 GHz. Then, SAP moves to a
2.4 GHz channel while LL_LT_SAP remains on
5 GHz. After this transition, SAP attempts
SCC with LL_LT_SAP.

To avoid SCC with LL_LT_SAP, it is necessary
to check if both frequencies are on the same
MAC and LL_LT_SAP is present. If these
conditions are met, SCC should be skipped.

Change-Id: Ic2083cca7290789eed455dfecaa34debc3b4b7b3
CRs-Fixed: 4107653