qcacld-3.0: Move SAP to user config freq after STA disconnects
If a SAP is moved to existing STA channel due to concurrency,
After STA is disconnected move SAP to user given config.

1) In start_bss once the bss is started then user given config
is updated with the freq which SAP was brought up
But in case of ap_ap_scc_override scenario, user given freq
will be overridden with the interfering channel freq.
This leads to updating the overridden channel in user config
in start_bss.
To avoid this, check if user given freq is different from the
current freq after start_bss and update user config if different.

2) policy_mgr_is_sap_restart_required_after_sta_disconnect is
either called for a single vdev or all the existing vdevs.
If the vdev_id is INVALID_VDEV_ID then it is run for all the
vdevs.
But as per current logic, it always breaks after first vdev and ignores
rest all vdev.
So use different variable to iterate all the vdevs.

Change-Id: Idc1eb53fec76b34649e64ac74eebd11095b5ae03
CRs-Fixed: 4099656
(cherry picked from commit c2fca6bafd6b5ce5b831348e3db73e491b783ebe)
CRs-Fixed: 4276558