qcacld-3.0: Skip CSA received STA from force SCC
SAP+SAP+STA concurrency with one of the SAP is
in SCC with the STA on a DFS channel. In this case
when STA receives CSA from peer AP with no Tx mode
and new freq is another dfs freq, SAP which is in
SCC with STA moves to a non-dfs 5 GHz channel. Post
completion of SAP CSA, concurrent interference check
gets triggered, which finds the non-DFS 5 GHz freq
of SAP as MCC with the STA. This leads to SAP again
moving to the DFS channel where STA received CSA
with no Tx. This issue happens only if multiple
beaconing interfaces are present.

To avoid this, skip the STA vdevs which
received CSA with no TX in the force SCC logic.

Change-Id: Ie168bf86010e96474184ee7517d0627a5c82a141
CRs-Fixed: 4095175
(cherry picked from commit 10911e5eddf49dd4adae75bf2cb2e5b036455d36)
CRs-Fixed: 4276692