qcacld-3.0: Send the correct reason code for chipset logging
When the get ring buffer data request comes for chipset logging then
WLAN_LOG_REASON_CODE_UNUSED reason code is sent to userspace, which
is not correct. The reason code should be according to the final
dump_in_progress value. If the final value is 1, then the chipset log
will be sent with the reason code WLAN_LOG_REASON_DUMP_IN_PROGRESS;
otherwise, the reason code remains the same as earlier.

Added checks to ensure the correct reason code is set based on the
dump_in_progress value.

CRs-Fixed: 4112965
Change-Id: I49e7d1d02975feb99d79f876460242112d53cc25