qcacld-3.0: Fix stats updation during sample/window transition
Currently there are the below three issues:
1) PKT IAT is not counted as 0 for the first packet in a sample.
2) Min-max stats when carried forward between two windows
of the same sample, are getting over-written.
3) If there is no packet tx/rx during window-2 of a sample
the min-max stats from window-1 are not carried forward to
the window-2, resulting in all the min-max stats being 0 for
the window-2.

The current change fixes the above issues in the below mentioned way
1) When the transition flags indicate sample transition,
the packet IAT is excluded from the min/max stats calculation
2) Use atomic transition flag setting to indicate window transition
from win-1 to win-2, which will be tested-and-cleared by the first
packet in the next window to cache the win-1 min/max stats to win-2
min/max stats, thereby carrying them forward to the 2nd window of the
sample. Atomic test-and-clear takes of the race condition.
3) When collecting the window-2 stats in the STC timer,
if the packets count has not incremented, then we go ahead
and pick the min/max stats from win-1 of the same sample.

Change-Id: I02b4bff55a724d6c7d6e1b0a309e9a98ac0850d5
CRs-Fixed: 4117726