qcacld-3.0: Skip invalid frequency while SAP pick up new channel
In the SAP + STA SCC on DFS channel user scenario:
1. SAP start on 5 GHz non-DFS frequency(5745) under US
2. Change country code to DE(5745 was disabled by user)
3. SAP do channel switch to 5180
4. STA connect to external AP on DFS channel 100
5. SAP follow STA with SCC and move to channel 100
6. external AP power off
7. STA happen beacon miss
8. SAP will pick up new channel after STA disconnect

Then in the step#8, SAP will pick up user_config_freq with
priority, but in this case, the user_config_freq is 5745
that is not valid for new country code, which will cause
SAP not move.

So in order to fix this issue, need to skip the invalid
frequency from pcl to make sure SAP move out from DFS
channel to new one.

Change-Id: I8ad3a734c23a703e7f2b4a334e934dead39cfbfc
CRs-Fixed: 4119349