qcacld-3.0: Allow LPC stop via vendor command
This change introduces an enhancement,to stop
LPC via vendor command and allow the device to
enter PM modes without disabling/removing the
monitor interface.

When the user sets the filter to zero,host will
stop LPC and release the wakelock. When valid filters
are applied again wakelock will be acquired and LPC
will be enabled again

Change-Id: If9a56502bb7907eb136293ea89d8d3d3e55a1a0b
CRs-Fixed: 4123013