qcacld-3.0: Handle case of flow tuple change with same socket
There can be a possible case, where the data from
a particular socket can be destined to a different
IP-PORT combo as compared to the IP-PORT combo with
which the data traffic started.

In the above mentioned case, since the socket has not
changed, the TX AFT does not get updated with the new
flow tuple and hence affects STC in shortlisting
bi-directional traffic.

In order to fix this issue, the packet flow info is
compared against the TX AFT flow info on a periodic
basis to ensure that the TX AFT flow info stays in
sync with the flow info of the actual data traffic.

Change-Id: If49625abfbb639b085ebb0e7276a04d2091c32bf
CRs-Fixed: 4125491