qcacld-3.0: Fix compilation warnings in following APIs
Fix compilation warnings in the current version of the
following APIs:
1. wlan_connectivity_disconnect_event() - The variable int_reason
is not initialized in the API and does not get a valid value
if reason < REASON_PROP_START.
2. policy_mgr_update_with_safe_channel_list() - The variable
nan_5g_freq is not initialized in the API and gets a valid
value only if NAN discovery is active.
3.sme_update_he_cap_nss(): In the for loop, the calls to the
macros HE_SET_MCS_FOR_NSS are statements with no effect.

To fix these compilation warnings:
1. Initialize the variables int_reason and nan_5g_freq to zero
in the APIs wlan_connectivity_disconnect_event and
policy_mgr_update_with_safe_channel_list.
2. Assign appropriate values to tx_mcs_map and rx_mcs_map
inside the for loop.

CRs-Fixed: 4123312
Change-Id: I9545cb1decd0207c6302bc948eb2a3e1671cd5ba