qcacld-3.0: Fix use-after-free issue in p2p scan abort
In the current implementation, the scan request is
freed in case of an error. However, this leads to issues when
the same request is accessed during the abort process of the
P2P scan.

To avoid issue Ensure that the scan request is not used if it
has already been freed.

Change-Id: I94e4061cb8bcf51f41a661c384e201fb36fa6e94
CRs-Fixed: 4120540