qcacld-3.0: Sample only TxRx stage for low packet rate flows
For voice traffic type, the packet rate is very low.
Hence to detect this flow type, reduce the packet rate and
restrict the tracking of this flow for TxRx stage only as
reducing the packet rate can cause lots of flows to get into
the sampling table and will stay in the table till last stats
report is not sent.
Set last_stats_report_ts for such low packet rate flows when
the TxRx statistics report is sent.
As this can cause lots of traffic to come into the sampling
table, increase the size of sampling table to 128 and give
3/4th quota to bi-directional flows, 1/4th quota to Rx flows
and remove quota for only Tx flows (UDP Tx flow) as they are
not being tracked currently.

Change-Id: Ide48464bd54e4884213e3c63d32c724d1ce73f91
CRs-Fixed: 4128909