qcacld-3.0: Abort ongoing scan when initiating Monitor mode in concurrency
During the Monitor mode concurrency scenario, due to on-going scan
the primary mac is configured for 2.4 GHz while secondary MAC
is configured for 5/6 GHz. When Monitor mode is initiated on 5/6 GHz
band, it is initiated on secondary MAC which is not support and it
leads to FW crash.

Abort ongoing scan when initiating Monitor mode in STA+MON
concurrency

Change-Id: I7b0736436473a544a54fc93612f127548762493f
CRs-Fixed: 4117770