qcacld-3.0: set TID for realtime flows
If a flow is classified as realtime traffic type,
which are, video call, voice call and gaming, set
the TID provided in the classified flow result.
To set TID, in sk_buff's mark field update ul_tid
with classified tag which will be retrieved and
stored in MSDU info and enqueued in tx ring.

Change-Id: Ided7e12f8e638d626c5eb82917445d11f55d522c
CRs-Fixed: 4113261