qcacld-3.0: Defer RSO_STOP when roaming is in progress
Currently, driver attempts to send RSO_STOP to fw when concurrent
SAP tries to move to STA channel and this is dropped if STA
roaming is still in progress. STA is at the end of roam completion
but states are not updated yet. SAP CSA waits till the roaming is
done and ROAM cmd is removed from SER.
So, avoid sending RSO_STOP to fw when roaming state is
still "in progress" and cache it to send once roaming is done.

Similarly, there might be a back to back RSO_STOP and RSO_START
as this is generic path. Check if there is a pending RSO_STOP
request and clear it upon RSO_START from the same requestor.

Change-Id: Id3cb199f19ff3b086923e67e4ee469f18a250beb
CRs-Fixed: 4124504