qcacld-3.0: Correct RX MC BC Count for SAP Big Data Stats
Currently, the host checks if the MSB is set for rx_mc_bc_cnt, as the
firmware sets the MSB for a valid value. However, the host can also
obtain rx_mc_bc_cnt from the DP, in which case the MSB will not be set.

To address this issue, reset the MSB from rx_mc_bc_cnt obtained from
the firmware. Additionally, remove the MSB set check when updating
rx_mc_bc_cnt to userspace as part of the big data stats.

CRs-Fixed: 4116773
Change-Id: I637504759aca8c2a9ef711c15ff33ae4c2a2b033