qcacld-3.0: Fix ML SAP CSA issue
Normally, SAP should not be MCC with STA. If CSA causes MCC on SAP +
STA, CSA should be rejected.

In the current logic for ML SAP, for the case of STA + SAP SCC on
Channel 36, if the user triggers CSA from Channel 36 to Channel 44, it
is allowed because the current state checking by
policy_mgr_is_scc_with_this_vdev_id is SCC on Channel 36 in the API
csr_check_concurrent_channel_overlap. This doesn't "break", and
intf_ch_freq is set to 0. Later, the CSA logic thinks there is no
channel overlap for the target channel and continues CSA, finally
causing MCC.

Fix: Add an additional check. If the target channel is different from
intf_ch_freq and on the same MAC always, then channel overlap happens
and CSA should be rejected.

Change-Id: I476da784bc5ffc8ac221852dfb6de679e2848f7e
CRs-Fixed: 4132950