qcacld-3.0: Restart ap_ecsa_timer for beacon interval in msec
When the CSA count is 10, the host sends WMI_BCN_TMPL_CMDID and
restarts the ap_ecsa_timer with a beacon interval of 100 TU upon
receiving the WMI_OFFLOAD_BCN_TX_STATUS_EVENT for CSA 10,
continuing this process until the CSA count reaches 0.

The firmware sends beacons at every TBTT interval of 102 msec.
Sometimes, there is a chances of receiving two WMI_BCN_TMPL_CMDID templates
in the firmware if the host sends them for every 100 TU.
Consequently, the CSA count will not be reflected in the beacons in
sequence, resulting in the failure of the following certification test
cases:

EHT-4.22.1_5G
EHT-4.22.1_24G
EHT-4.22.1_6G

To address the issue, Convert the beacon interval from TU to MS.

Change-Id: I26ba4157718b802c3a0ffd1ba24c43ea835477c6
CRs-Fixed: 4200808