qcacld-3.0: Update the channel width if EHT enabled
This concerns the DFS No Wait feature. It does not update the channel
width in populate_dot11f_vht_operation for cases where EHT is enabled.
As a result, the reference station connects to the DUT with a higher
bandwidth when the DUT is in the DFS No Wait CAC state.

Change-Id: I20a3d0b42dee273b096fcc19b760882ceadc8711
CRs-Fixed: 4140085