qcacld-3.0: Add support for cmd SET_RECONNECT_TO_SAME_BSSID
Add support for vendor command SET_RECONNECT_TO_SAME_BSSID
to set the value of time_to_enable_reconnection_again.
This value is used as the duration in seconds for which STA
is not allowed to select a target AP that has transmitted a
deauthentication frame or disassociation frame to avoid
repeated roaming and disconnections.

CRs-Fixed: 4141099
Change-Id: Icb5f273d32bc135f8820515bea6d923a00458b67