qcacld-3.0: Send VLP and C2C TPC for CSA while operating in C2C power
When SAP/STA operates in the 6 GHz C2C power type and
CSA is triggered for SAP, the host currently sends only
the C2C TPC to the firmware for both SAP and STA(for CSA
event received from SAP). However, the firmware needs to
cache the VLP TPC to use in the suspend state.
Add logic to send the VLP TPC before configuring the C2C
TPC to the firmware for both SAP and STA for CSA.

Change-Id: I846dfa399b6666c85613d19db26d1c80b9e98883
CRs-Fixed: 4138975