qcacld-3.0: Limit the size of vendor IEs from userspace
For large management frames support feature, if the length
of the vendor IEs exceed 1024 bytes, then the scan start cmd
from userspace exceeds in size. This results in scan failures.

To fix this, limit the size of the additional IEs to be added
in scan probes or in any management frames sent by host to 1024
bytes.

Change-Id: Ie48f4fe24f88ca580bbada6bc2f6e017d4f42dd5
CRs-Fixed: 4138914