qcacld-3.0: Don't use cached tx power after setting tx power
With the following sequence, step 4 returns wrong tx power as
the result of rate limit:
	iw wlan0 set txpower fixed 12
	iw dev|grep txpower # return 12db
	iw wlan0 set txpower fixed 7
	iw dev|grep txpower # return 12db

To fix above issue, add a new flag to indicate if tx power
in driver cache is valid. When not valid, get tx power from
firmware directly even if rate limit condition is met.

Change-Id: Iedeaa94e97cdee651edeae387c2eee6f143e8522
CRs-Fixed: 4099978