qcacld-3.0: Copy EHT OP to peer assoc to enable MCS15
To enable MCS15, F/W expects MCS15 BIT in EHT OP is set in peer assoc cmd,
host gets EHT OP from assoc rsp, and send to F/W directly.

As 802.11be_D7.0 EHT Operation Parameters field format, the MCS15 Disable
subfield(BIT6) indicates whether the reception of an EHT PPDU with
EHT-MCS 15 in both the Data field and EHT-SIG field by the AP is disabled
or enabled. The MCS15 Disable subfield is set to 1 to indicate that it is
disabled and set to 0 to indicate that it is enabled.

Change-Id: I4091f83bbe45f59ad52889f090f3c4234e7ba05d
CRs-Fixed: 4143358