qcacld-3.0: Add mcs15_disable to EHT OP IE
Add mcs15_disable generation and parsing support for EHT OP.

As 802.11be_D7.0 EHT Operation Parameters field format, the MCS15 Disable
subfield(BIT6) indicates whether the reception of an EHT PPDU with
EHT-MCS 15 in both the Data field and EHT-SIG field by the AP is disabled
or enabled. The MCS15 Disable subfield is set to 1 to indicate that it is
disabled and set to 0 to indicate that it is enabled.

Change-Id: Iaff3961fd074a8c65cc93f76717bdf64b10b72f0
CRs-Fixed: 4144218