qcacld-3.0: Relax TX flow eviction criteria for STC
Currently the TX flow is evicted based on the data rate
or inactivity criteria. This is fine in cases where the
presence of TX & RX flows does not interfere with any use-case.
But with STC running and attempting to classify flows,
the detection of bi-directional flows is very important.

Hence to achieve the above mentioned objective for STC,
do not evict the TX flow if a corresponding RX flow is
still present in the RX flow table. This will aid STC in
proper flow classification.

Change-Id: Id44a894271b17d2cb9acb1918572156bceff8a3b
CRs-Fixed: 4137036