qcacld-3.0: Fix invalid RSSI sent over the DISCONN event
The RSSI value sent for disconnection event is fetched from the
hdd link info. This information is stale in disconnection
scenario so RSSI value sent to the userspace is not in parity
with the RSSI observed in recent RX frames.

Fetch the RSSI value which was cached from the CP stats lost
link info event and update it in the the DISCONN connectivity
logging event.

CRs-Fixed: 4138116
Change-Id: I50a95d52cfba40ec069d3960b3b443b65ac44fa8