qcacld-3.0: Fixed T2LM Request Band info
Currently, the WMI management frame TX completion path does not
populate the frequency field. As a result, the band is derived
from an invalid frequency, leading to incorrect band information
in the T2LM Request.

Update the logic to derive and populate the band directly
instead of relying on the frequency. This avoids incorrect band
assignments in the T2LM Request and ensures consistency in the
logged data.

CRs-Fixed: 4136867
Change-Id: Ia614a002abf4b0a460d59ce9c3a9f79a843b6593