qcacld-3.0: Enable NAN Support on DFS Channels
NAN is disabled for channels marked as DFS in the channel list. This
logic ensures NAN does not operate on DFS frequencies.

Add support to enable NAN on DFS channels based on a new ini flag and
firmware capability. If the firmware advertises support for NAN over DFS
and the ini is enabled, NAN is allowed on DFS channels.

CRs-Fixed: 4150394
Change-Id: I706a58dae08384c087f9c85c5bd12fabcd0cf3cd