qcacld-3.0: Fix potential OOB access in STC candidate sampling table
As part of STC flow work handler, rx and bidirectional
flows information are are populated initially in candidate
sampling table in STC context which is of size 128. This
would be done in a loop based on rx fst size which is 256.
The candidate_idx variable is incremented in the loop
every time irrespective of whether the sampling candidate
table is populated or not. This would result in an OOB
access as candidate_idx could go beyond value of 128

Fix is to update candidate_idx only when sampling candidate
table is populated.

Change-Id: I6de7e17defa832cb091df61727b7196e3dc76329
CRs-Fixed: 4146014