qcacld-3.0: Add support for the LFR2.0 roaming between pmf APs
The current host driver does not support the LFR2.0 roaming if
some APs enabled pmf. In this case, if the previous AP which
out STA roamed/connected before but has not disconnect to our
STA, then if our STA roam back to this AP, it will send the
reassoc resp with reason 30(Association request rejected
temporarily, try again later), then STA will wait the time
interval(e.g 1 second), and then re-roam to this AP.

In the current design, host driver only re-send reassoc req
after the time interval, but some APs do not support to
receive reassoc req in this case, they need the STA re-send
the auth at first, and then re-send the reassoc req. Do add
this support in host driver to make the roam success in this
case.

Change-Id: Ib82be7926c2547b7594803f73fe6af311d90f7f4
CRs-Fixed: 4151215