qcacld-3.0: Avoid triggering bug report during STA deletion
Currently, cds_flush_logs() is triggered before the
disconnect indication is sent to userspace when a client
disconnects from SAP. This initiates a bug report and
causes the system to fetch logs via get_ring_data,
delaying the disconnect processing in hostapd.

During this delay, the same client may reconnect,
and the driver sends a new CONNECT indication.
Since hostapd hasn’t yet processed the earlier DISCONNECT,
it may treat the delayed DEL_STA as a command to remove the
newly connected client, leading to an unintended disconnection
and state mismatch between hostapd and the driver.

To prevent this, the bug report trigger is removed,
allowing hostapd to process the disconnect promptly
and avoid tearing down valid connections.

Change-Id: I69878c56bddaa4830c721dca8a0ae70125b06bbd
CRs-Fixed: 4157532