qcacld-3.0: Fix FISA timestamp update for flows when WLM is active
Currently when WLM is active, the packets are not aggregated in
FISA and the last_active timestamp is not updated as well.
This causes STC to detect an inactivity in this flow and send
flow_pause indication, even when the flow is active.

To fix this issue, we update the last_active timestamp in
FISA flow entry, even when aggregation is not done due to WLM.

Change-Id: I95dbf603c0304f88c3d759e25ac40ec78c2d7353
CRs-Fixed: 4157362