qcacld-3.0: Fix the classified state of Tx-Rx flows
Once STC completes the classification of the Tx-Rx flows,
all these Tx-Rx flows are marked as classified and the
corresponding classified_flow_idx is updated in the Tx-Rx
flows. There is no differentiation done between flows
classified as known or unknown. Since the classified_flow_idx
is not updated for flows which are not classified into
known traffic types, their deletion may lead to removal
of the classified_flow at index 0 in the classified flow
table.

In order to fix the above mentioned issue, we introduce
a state for the classified flag in the Tx-Rx flows, which
can be used for decision making when removing the Tx-Rx
flows and purging the classified flow table entry.

Change-Id: I97bb6b661cdf52cb30653e4407ed809155eeea4c
CRs-Fixed: 4160700