qcacld-3.0: Fix wrong BW indication in HE capability IE
The bandwidth field(chan_width_1 = 40 & 80 MHz in the 5 GHz band)
not get updated properly in the 5 GHz SAP 20 MHz switch to
5 GHz SAP 80 MHz, so remove the bandwidth limitation to fix
wrong BW indication issue.

eg:
1. SAP DUT start on channel 165 20 MHz
2. STA DUT connect to external AP on channel 149
3. SAP DUT do SCC and switch to 149 with BW=80 MHz

Then the "40 & 80 MHz in the 5 GHz band" field in the HE Capability
IE of beacon/probe response still is zero not change to 1.

Change-Id: Ia22375992ba5c9f99b523772537b20489f3d1c19
CRs-Fixed: 4162982